Create codacy-analysis.yaml
Analyzes each commit or pull request by running all supported static code analysis tools for the languages found in your repository. 
Prints the analysis results on the console, which is visible on the GitHub Action's workflow panel. 
Fails the workflow if it finds at least one issue in your code.